+++ client/idrepo/console/src/main/java/org/apache/syncope/client/console/rest/ReportRestClient.java
@@ -116,7 +116,7 @@ public class ReportRestClient extends BaseRestClient implements ExecutionRestCli
     }
 
     @Override
-    public int countExecutions(final String taskKey) {
+    public long countExecutions(final String taskKey) {
         return getService(ReportService.class).
                 listExecutions(new ExecQuery.Builder().key(taskKey).page(1).size(0).build()).getTotalCount();
     }
