+++ client/idrepo/console/src/main/java/org/apache/syncope/client/console/rest/TaskRestClient.java
@@ -64,23 +64,23 @@ public class TaskRestClient extends BaseRestClient implements ExecutionRestClien
         getService(TaskService.class).actionJob(refKey, jobAction);
     }
 
-    public int count(final TaskType kind) {
+    public long count(final TaskType kind) {
         return getService(TaskService.class).search(
                 new TaskQuery.Builder(kind).page(1).size(0).build()).getTotalCount();
     }
 
-    public int count(final String resource, final TaskType kind) {
+    public long count(final String resource, final TaskType kind) {
         return getService(TaskService.class).search(
                 new TaskQuery.Builder(kind).resource(resource).page(1).size(0).build()).getTotalCount();
     }
 
-    public int count(final AnyTypeKind anyTypeKind, final String entityKey, final TaskType kind) {
+    public long count(final AnyTypeKind anyTypeKind, final String entityKey, final TaskType kind) {
         return getService(TaskService.class).search(
                 new TaskQuery.Builder(kind).anyTypeKind(anyTypeKind).entityKey(entityKey).page(1).size(0).build()).
                 getTotalCount();
     }
 
-    public int count(final AnyTypeKind anyTypeKind, final String entityKey, final String notification) {
+    public long count(final AnyTypeKind anyTypeKind, final String entityKey, final String notification) {
         return getService(TaskService.class).search(
                 new TaskQuery.Builder(TaskType.NOTIFICATION).notification(notification).
                         anyTypeKind(anyTypeKind).entityKey(entityKey).page(1).size(0).build()).
@@ -88,7 +88,7 @@ public class TaskRestClient extends BaseRestClient implements ExecutionRestClien
     }
 
     @Override
-    public int countExecutions(final String taskKey) {
+    public long countExecutions(final String taskKey) {
         return getService(TaskService.class).
                 listExecutions(new ExecQuery.Builder().key(taskKey).page(1).size(0).build()).getTotalCount();
     }
