+++ client/idrepo/console/src/main/java/org/apache/syncope/client/console/rest/UserRestClient.java
@@ -69,7 +69,7 @@ public class UserRestClient extends AbstractAnyRestClient<UserTO> {
     }
 
     @Override
-    public int count(final String realm, final String fiql, final String type) {
+    public long count(final String realm, final String fiql, final String type) {
         return getService(UserService.class).
                 search(new AnyQuery.Builder().realm(realm).fiql(fiql).page(1).size(0).details(false).build()).
                 getTotalCount();
